Rename error.rb to errors.rb
