+++ lib/hanami/router/errors.rb
